Refactor drs tests into separate file
